refactor(autoware_tensorrt_yolox): move TrtYoloXDetector to its own files
Move the TrtYoloXDetector class (and TrtYoloXDetectorConfig /
TrtYoloXDetectorResult) out of tensorrt_yolox_node.hpp/.cpp into dedicated
tensorrt_yolox_detector.hpp/.cpp, and register the new source in CMakeLists.
This is a pure relocation with no behavior change; tensorrt_yolox_node.hpp now
includes tensorrt_yolox_detector.hpp and the node-only includes are trimmed.
