+++ perception/autoware_tensorrt_yolox/CMakeLists.txt
@@ -150,6 +150,7 @@ rclcpp_components_register_node(yolox_single_image_inference_node
 
 ament_auto_add_library(${PROJECT_NAME}_node SHARED
   src/tensorrt_yolox_node.cpp
+  src/tensorrt_yolox_detector.cpp
 )
 
 ament_target_dependencies(${PROJECT_NAME}_node
